+++ src/features/searchForm/__tests__/useSearchForm.test.ts
@@ -1,5 +1,6 @@
 import { renderHook, act } from '@testing-library/react';
 import useSearchForm from '../useSearchForm';
+import { createInputChangeEvent, createFormSubmitEvent } from '@utils/tests/events';
 
 // mock useNavigate
 const mockNavigate = vi.fn();
@@ -13,38 +14,38 @@ vi.mock('react-router-dom', async () => {
 
 describe('useSearchForm hook', () => {
   test('updates search term on handleFormChange', () => {
-    const { result } = renderHook(() => useSearchForm());
+    const { result } = renderHook(() => useSearchForm('term', 'page'));
 
     act(() => {
-      result.current.handleFormChange({ target: { value: 'term' } });
+      result.current.handleFormChange(createInputChangeEvent('search', 'term'));
     });
 
     expect(result.current.searchFormTerm).toBe('term');
   });
 
   test('updates error for missing search term', () => {
-    const { result } = renderHook(() => useSearchForm());
+    const { result } = renderHook(() => useSearchForm('', 'page'));
 
     act(() => {
-      result.current.handleFormChange({ target: { value: '' } });
+      result.current.handleFormChange(createInputChangeEvent('search', ''));
     });
 
     act(() => {
-      result.current.handleFormSubmit({ preventDefault: vi.fn() });
+      result.current.handleFormSubmit(createFormSubmitEvent());
     });
 
     expect(result.current.searchFormError).toBe('Please enter a search term');
   });
 
   test('navigates to search page for valid form submission', () => {
-    const { result } = renderHook(() => useSearchForm());
+    const { result } = renderHook(() => useSearchForm('term', 'page'));
 
     act(() => {
-      result.current.handleFormChange({ target: { value: 'term' } });
+      result.current.handleFormChange(createInputChangeEvent('search', 'term'));
     });
 
     act(() => {
-      result.current.handleFormSubmit({ preventDefault: vi.fn() });
+      result.current.handleFormSubmit(createFormSubmitEvent());
     });
 
     expect(mockNavigate).toHaveBeenCalledWith('/search/term');